feat: expose Event deserialization via ParseEvent
