+++ event.go
@@ -8,9 +8,23 @@ import (
 	"bufio"
 	"bytes"
 	"context"
+	"encoding/base64"
+	"errors"
+	"fmt"
 	"io"
 	"time"
 )
+
+var (
+	headerID    = []byte("id:")
+	headerData  = []byte("data:")
+	headerEvent = []byte("event:")
+	headerRetry = []byte("retry:")
+)
+
+type EventParseConfig struct {
+	EncodingBase64 bool
+}
 
 // Event holds all of the event source fields
 type Event struct {
@@ -20,6 +34,67 @@ type Event struct {
 	Event     []byte
 	Retry     []byte
 	Comment   []byte
+}
+
+func ParseEvent(msg []byte, cfg EventParseConfig) (*Event, error) {
+	var e Event
+	var err error
+
+	if len(msg) < 1 {
+		return nil, errors.New("event message was empty")
+	}
+
+	// Normalize the crlf to lf to make it easier to split the lines.
+	// Split the line by "\n" or "\r", per the spec.
+	for _, line := range bytes.FieldsFunc(msg, func(r rune) bool { return r == '\n' || r == '\r' }) {
+		switch {
+		case bytes.HasPrefix(line, headerID):
+			e.ID = append([]byte(nil), trimHeader(len(headerID), line)...)
+		case bytes.HasPrefix(line, headerData):
+			// The spec allows for multiple data fields per event, concatenated them with "\n".
+			e.Data = append(e.Data[:], append(trimHeader(len(headerData), line), byte('\n'))...)
+		// The spec says that a line that simply contains the string "data" should be treated as a data field with an empty body.
+		case bytes.Equal(line, bytes.TrimSuffix(headerData, []byte(":"))):
+			e.Data = append(e.Data, byte('\n'))
+		case bytes.HasPrefix(line, headerEvent):
+			e.Event = append([]byte(nil), trimHeader(len(headerEvent), line)...)
+		case bytes.HasPrefix(line, headerRetry):
+			e.Retry = append([]byte(nil), trimHeader(len(headerRetry), line)...)
+		default:
+			// Ignore any garbage that doesn't match what we're looking for.
+		}
+	}
+
+	// Trim the last "\n" per the spec.
+	e.Data = bytes.TrimSuffix(e.Data, []byte("\n"))
+
+	if cfg.EncodingBase64 {
+		buf := make([]byte, base64.StdEncoding.DecodedLen(len(e.Data)))
+
+		n, decodeErr := base64.StdEncoding.Decode(buf, e.Data)
+		if decodeErr != nil {
+			err = fmt.Errorf("failed to decode event message: %s", decodeErr)
+		}
+		e.Data = buf[:n]
+	}
+	return &e, err
+}
+
+func trimHeader(size int, data []byte) []byte {
+	if data == nil || len(data) < size {
+		return data
+	}
+
+	data = data[size:]
+	// Remove optional leading whitespace
+	if len(data) > 0 && data[0] == 32 {
+		data = data[1:]
+	}
+	// Remove trailing new line
+	if len(data) > 0 && data[len(data)-1] == 10 {
+		data = data[:len(data)-1]
+	}
+	return data
 }
 
 func (e *Event) hasContent() bool {
